bug fix on the flatness demo + car footprint module from another repo
The bug:
==========================
My flatness-based trajectory generation demo was designed such that
the trajectory of the car's kinematic center (i.e. X,Y)]
is the same whether driving in forward / backward.
This means the initial yaw angle needs to be taken into account
(which was overlooked)!

How was the bug discovered?
==========================
I discovered it when I tried to visualize the footprint
(with the car driving in reverse).
This is why both seemingly unrelated changes
appear in the same commit.

btw some enhancements to the notebook
==========================

* (needless to say) the addition of footprint visualization
  (to make the visualization easier, the XY-path is enlarged,
   which also means that the steering angle becomes more realistic.)
* better self-documentation, especially the paragraphs on
  * "what to try" and
  * the last paragraph --- more remarks that are relevant in practice.